+++ packages/snaps-controllers/src/snaps/registry/SnapsRegistryController.test.ts
@@ -7,9 +7,9 @@ import {
 import type { SemVerRange, SemVerVersion } from '@metamask/utils';
 import fetchMock from 'jest-fetch-mock';
 
-import type { JsonSnapsRegistryArgs } from './json';
-import { JsonSnapsRegistry } from './json';
 import { SnapsRegistryStatus } from './registry';
+import type { SnapsRegistryControllerArgs } from './SnapsRegistryController';
+import { SnapsRegistryController } from './SnapsRegistryController';
 import { getRestrictedSnapsRegistryControllerMessenger } from '../../test-utils';
 
 // Public key for the private key:
@@ -17,10 +17,10 @@ import { getRestrictedSnapsRegistryControllerMessenger } from '../../test-utils'
 const MOCK_PUBLIC_KEY =
   '0x03a885324b8520fba54a173999629952cfa1f97930c20902ec389f9c32c6ffbc40';
 
-const getRegistry = (args?: Partial<JsonSnapsRegistryArgs>) => {
+const getRegistry = (args?: Partial<SnapsRegistryControllerArgs>) => {
   const messenger = getRestrictedSnapsRegistryControllerMessenger();
   return {
-    registry: new JsonSnapsRegistry({
+    registry: new SnapsRegistryController({
       messenger,
       publicKey: MOCK_PUBLIC_KEY,
       clientConfig: {
@@ -125,7 +125,7 @@ const MOCK_COMPATIBILITY_SIGNATURE_FILE = {
   format: 'DER',
 };
 
-describe('JsonSnapsRegistry', () => {
+describe('SnapsRegistryController', () => {
   fetchMock.enableMocks();
 
   afterEach(() => {
@@ -138,7 +138,7 @@ describe('JsonSnapsRegistry', () => {
       .mockResponseOnce(JSON.stringify(MOCK_SIGNATURE_FILE));
 
     const { messenger } = getRegistry();
-    const result = await messenger.call('SnapsRegistry:get', {
+    const result = await messenger.call('SnapsRegistryController:getSnap', {
       [MOCK_SNAP_ID]: {
         version: '1.0.0' as SemVerVersion,
         checksum: DEFAULT_SNAP_SHASUM,
@@ -159,7 +159,7 @@ describe('JsonSnapsRegistry', () => {
       .mockResponseOnce(JSON.stringify(MOCK_EMPTY_SIGNATURE_FILE));
 
     const { messenger } = getRegistry();
-    const result = await messenger.call('SnapsRegistry:get', {
+    const result = await messenger.call('SnapsRegistryController:getSnap', {
       [MOCK_SNAP_ID]: {
         version: '1.0.0' as SemVerVersion,
         checksum: DEFAULT_SNAP_SHASUM,
@@ -179,7 +179,7 @@ describe('JsonSnapsRegistry', () => {
       .mockResponseOnce(JSON.stringify(MOCK_SIGNATURE_FILE));
 
     const { messenger } = getRegistry();
-    const result = await messenger.call('SnapsRegistry:get', {
+    const result = await messenger.call('SnapsRegistryController:getSnap', {
       [MOCK_SNAP_ID]: {
         version: '1.0.1' as SemVerVersion,
         checksum: DEFAULT_SNAP_SHASUM,
@@ -199,7 +199,7 @@ describe('JsonSnapsRegistry', () => {
       .mockResponseOnce(JSON.stringify(MOCK_SIGNATURE_FILE));
 
     const { messenger } = getRegistry();
-    const result = await messenger.call('SnapsRegistry:get', {
+    const result = await messenger.call('SnapsRegistryController:getSnap', {
       [MOCK_SNAP_ID]: {
         version: '1.0.0' as SemVerVersion,
         checksum: 'bar',
@@ -219,7 +219,7 @@ describe('JsonSnapsRegistry', () => {
       .mockResponseOnce(JSON.stringify(MOCK_SIGNATURE_FILE));
 
     const { messenger } = getRegistry();
-    const result = await messenger.call('SnapsRegistry:get', {
+    const result = await messenger.call('SnapsRegistryController:getSnap', {
       [MOCK_SNAP_ID]: {
         version: '1.0.0' as SemVerVersion,
         checksum: 'foo',
@@ -240,7 +240,7 @@ describe('JsonSnapsRegistry', () => {
       .mockResponseOnce(JSON.stringify(MOCK_SIGNATURE_FILE));
 
     const { messenger } = getRegistry();
-    const result = await messenger.call('SnapsRegistry:get', {
+    const result = await messenger.call('SnapsRegistryController:getSnap', {
       'npm:@consensys/starknet-snap': {
         version: '0.1.10' as SemVerVersion,
         checksum: DEFAULT_SNAP_SHASUM,
@@ -267,7 +267,7 @@ describe('JsonSnapsRegistry', () => {
         database: { verifiedSnaps: {}, blockedSnaps: [] },
       },
     });
-    const result = await messenger.call('SnapsRegistry:get', {
+    const result = await messenger.call('SnapsRegistryController:getSnap', {
       [MOCK_SNAP_ID]: {
         version: '1.0.0' as SemVerVersion,
         checksum: DEFAULT_SNAP_SHASUM,
@@ -287,7 +287,7 @@ describe('JsonSnapsRegistry', () => {
       .mockResponseOnce(JSON.stringify(MOCK_COMPATIBILITY_SIGNATURE_FILE));
 
     const { messenger } = getRegistry();
-    const result = await messenger.call('SnapsRegistry:get', {
+    const result = await messenger.call('SnapsRegistryController:getSnap', {
       [MOCK_SNAP_ID]: {
         version: '1.0.0' as SemVerVersion,
         checksum: DEFAULT_SNAP_SHASUM,
@@ -307,7 +307,7 @@ describe('JsonSnapsRegistry', () => {
       .mockResponseOnce(JSON.stringify(MOCK_COMPATIBILITY_SIGNATURE_FILE));
 
     const { messenger } = getRegistry();
-    const result = await messenger.call('SnapsRegistry:get', {
+    const result = await messenger.call('SnapsRegistryController:getSnap', {
       [MOCK_SNAP_ID]: {
         version: '1.1.0' as SemVerVersion,
         checksum: DEFAULT_SNAP_SHASUM,
@@ -332,7 +332,7 @@ describe('JsonSnapsRegistry', () => {
       },
     });
 
-    const result = await messenger.call('SnapsRegistry:get', {
+    const result = await messenger.call('SnapsRegistryController:getSnap', {
       [MOCK_SNAP_ID]: {
         version: '1.0.0' as SemVerVersion,
         checksum: DEFAULT_SNAP_SHASUM,
@@ -357,7 +357,7 @@ describe('JsonSnapsRegistry', () => {
       },
     });
 
-    const result = await messenger.call('SnapsRegistry:get', {
+    const result = await messenger.call('SnapsRegistryController:getSnap', {
       [MOCK_SNAP_ID]: {
         version: '1.0.1' as SemVerVersion,
         checksum: DEFAULT_SNAP_SHASUM,
@@ -376,7 +376,7 @@ describe('JsonSnapsRegistry', () => {
 
     const { messenger } = getRegistry();
 
-    const result = await messenger.call('SnapsRegistry:get', {
+    const result = await messenger.call('SnapsRegistryController:getSnap', {
       [MOCK_SNAP_ID]: {
         version: '1.0.0' as SemVerVersion,
         checksum: DEFAULT_SNAP_SHASUM,
@@ -399,7 +399,7 @@ describe('JsonSnapsRegistry', () => {
 
     const { messenger } = getRegistry();
 
-    const result = await messenger.call('SnapsRegistry:get', {
+    const result = await messenger.call('SnapsRegistryController:getSnap', {
       [MOCK_SNAP_ID]: {
         version: '1.0.0' as SemVerVersion,
         checksum: DEFAULT_SNAP_SHASUM,
@@ -423,7 +423,7 @@ describe('JsonSnapsRegistry', () => {
         '0x034ca27b046507d1a9997bddc991b56d96b93d4adac3a96dfe01ce450bfb661455',
     });
 
-    const result = await messenger.call('SnapsRegistry:get', {
+    const result = await messenger.call('SnapsRegistryController:getSnap', {
       [MOCK_SNAP_ID]: {
         version: '1.0.0' as SemVerVersion,
         checksum: DEFAULT_SNAP_SHASUM,
@@ -445,7 +445,7 @@ describe('JsonSnapsRegistry', () => {
 
       const { messenger } = getRegistry();
       const result = await messenger.call(
-        'SnapsRegistry:resolveVersion',
+        'SnapsRegistryController:resolveSnapVersion',
         MOCK_SNAP_ID,
         '^1.0.0' as SemVerRange,
       );
@@ -460,7 +460,7 @@ describe('JsonSnapsRegistry', () => {
 
       const { messenger } = getRegistry();
       const result = await messenger.call(
-        'SnapsRegistry:resolveVersion',
+        'SnapsRegistryController:resolveSnapVersion',
         MOCK_SNAP_ID,
         '^1.0.0' as SemVerRange,
       );
@@ -477,7 +477,7 @@ describe('JsonSnapsRegistry', () => {
         clientConfig: { type: 'extension', version: '15.0.0' as SemVerVersion },
       });
       const result = await messenger.call(
-        'SnapsRegistry:resolveVersion',
+        'SnapsRegistryController:resolveSnapVersion',
         MOCK_SNAP_ID,
         '^1.0.0' as SemVerRange,
       );
@@ -496,7 +496,7 @@ describe('JsonSnapsRegistry', () => {
       const { messenger } = getRegistry();
       expect(
         await messenger.call(
-          'SnapsRegistry:resolveVersion',
+          'SnapsRegistryController:resolveSnapVersion',
           MOCK_SNAP_ID,
           range,
         ),
@@ -512,7 +512,7 @@ describe('JsonSnapsRegistry', () => {
       const { messenger } = getRegistry();
       expect(
         await messenger.call(
-          'SnapsRegistry:resolveVersion',
+          'SnapsRegistryController:resolveSnapVersion',
           MOCK_SNAP_ID,
           range,
         ),
@@ -532,7 +532,7 @@ describe('JsonSnapsRegistry', () => {
         },
       });
       const result = await messenger.call(
-        'SnapsRegistry:resolveVersion',
+        'SnapsRegistryController:resolveSnapVersion',
         MOCK_SNAP_ID,
         '^1.0.0' as SemVerRange,
       );
@@ -568,7 +568,7 @@ describe('JsonSnapsRegistry', () => {
         },
       });
       const result = await messenger.call(
-        'SnapsRegistry:resolveVersion',
+        'SnapsRegistryController:resolveSnapVersion',
         MOCK_SNAP_ID,
         '^1.0.0' as SemVerRange,
       );
@@ -584,8 +584,11 @@ describe('JsonSnapsRegistry', () => {
         .mockResponseOnce(JSON.stringify(MOCK_SIGNATURE_FILE));
 
       const { messenger } = getRegistry();
-      await messenger.call('SnapsRegistry:update');
-      const result = messenger.call('SnapsRegistry:getMetadata', MOCK_SNAP_ID);
+      await messenger.call('SnapsRegistryController:updateRegistry');
+      const result = messenger.call(
+        'SnapsRegistryController:getSnapMetadata',
+        MOCK_SNAP_ID,
+      );
 
       expect(result).toStrictEqual({
         name: 'Mock Snap',
@@ -598,8 +601,11 @@ describe('JsonSnapsRegistry', () => {
         .mockResponseOnce(JSON.stringify(MOCK_SIGNATURE_FILE));
 
       const { messenger } = getRegistry();
-      await messenger.call('SnapsRegistry:update');
-      const result = messenger.call('SnapsRegistry:getMetadata', 'foo');
+      await messenger.call('SnapsRegistryController:updateRegistry');
+      const result = messenger.call(
+        'SnapsRegistryController:getSnapMetadata',
+        'foo',
+      );
 
       expect(result).toBeNull();
     });
@@ -612,7 +618,7 @@ describe('JsonSnapsRegistry', () => {
         .mockResponseOnce(JSON.stringify(MOCK_SIGNATURE_FILE));
 
       const { messenger } = getRegistry();
-      await messenger.call('SnapsRegistry:update');
+      await messenger.call('SnapsRegistryController:updateRegistry');
 
       expect(fetchMock).toHaveBeenCalledTimes(2);
     });
@@ -632,7 +638,7 @@ describe('JsonSnapsRegistry', () => {
           databaseUnavailable: false,
         },
       });
-      await messenger.call('SnapsRegistry:update');
+      await messenger.call('SnapsRegistryController:updateRegistry');
 
       expect(fetchMock).toHaveBeenCalledTimes(2);
       expect(spy).not.toHaveBeenCalled();
@@ -644,8 +650,8 @@ describe('JsonSnapsRegistry', () => {
         .mockResponseOnce(JSON.stringify(MOCK_SIGNATURE_FILE));
 
       const { messenger } = getRegistry();
-      await messenger.call('SnapsRegistry:update');
-      await messenger.call('SnapsRegistry:update');
+      await messenger.call('SnapsRegistryController:updateRegistry');
+      await messenger.call('SnapsRegistryController:updateRegistry');
 
       expect(fetchMock).toHaveBeenCalledTimes(2);
     });
@@ -657,8 +663,8 @@ describe('JsonSnapsRegistry', () => {
 
       const { messenger } = getRegistry();
       await Promise.all([
-        messenger.call('SnapsRegistry:update'),
-        messenger.call('SnapsRegistry:update'),
+        messenger.call('SnapsRegistryController:updateRegistry'),
+        messenger.call('SnapsRegistryController:updateRegistry'),
       ]);
 
       expect(fetchMock).toHaveBeenCalledTimes(2);
@@ -670,7 +676,11 @@ describe('JsonSnapsRegistry', () => {
       const { registry } = getRegistry();
 
       expect(
-        deriveStateFromMetadata(registry.state, registry.metadata, 'anonymous'),
+        deriveStateFromMetadata(
+          registry.state,
+          registry.metadata,
+          'includeInDebugSnapshot',
+        ),
       ).toMatchInlineSnapshot(`{}`);
     });