+++ packages/snaps-controllers/src/snaps/registry/index.ts
@@ -1,2 +1,22 @@
-export * from './registry';
-export * from './json';
+export type {
+  SnapsRegistryInfo,
+  SnapsRegistryMetadata,
+  SnapsRegistryRequest,
+  SnapsRegistryResult,
+} from './registry';
+export { SnapsRegistryStatus } from './registry';
+export type {
+  SnapsRegistryActions,
+  SnapsRegistryControllerArgs,
+  SnapsRegistryGetStateAction,
+  SnapsRegistryMessenger,
+  SnapsRegistryState,
+  SnapsRegistryStateChangeEvent,
+} from './SnapsRegistryController';
+export { SnapsRegistryController } from './SnapsRegistryController';
+export type {
+  SnapsRegistryControllerGetSnapAction,
+  SnapsRegistryControllerGetSnapMetadataAction,
+  SnapsRegistryControllerResolveSnapVersionAction,
+  SnapsRegistryControllerUpdateRegistryAction,
+} from './SnapsRegistryController-method-action-types';