+++ packages/snaps-controllers/CHANGELOG.md
@@ -56,6 +56,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `HandleRequest` is now `ExecutionServiceHandleRequestAction`.
     - `TerminateSnap` is now `ExecutionServiceTerminateSnapAction`.
     - `GetExecutionStatus` is now `ExecutionServiceGetExecutionStatusAction`.
+  - `SnapsRegistryController` actions:
+    - `GetResult` is now `SnapsRegistryControllerGetSnapAction`.
+      - Note: The method is now called `getSnap` instead of `get`.
+    - `GetMetadata` is now `SnapsRegistryControllerGetSnapMetadataAction`.
+      - Note: The method is now called `getSnapMetadata` instead of `getMetadata`.
+    - `ResolveVersion` is now `SnapsRegistryControllerResolveSnapVersionAction`.
+      - Note: The method is now called `resolveSnapVersion` instead of `resolveVersion`.
+    - `Update` is now `SnapsRegistryControllerUpdateRegistryAction`.
+      - Note: The method is now called `updateRegistry` instead of `update`.
 - **BREAKING:** All event types were renamed from `OnSomething` to `ControllerOnSomethingEvent` ([#3907](https://github.com/MetaMask/snaps/pull/3907), [#3916](https://github.com/MetaMask/snaps/pull/3916))
   - `SnapController` events:
     - `SnapStateChange` was removed in favour of `SnapControllerStateChangeEvent`.
@@ -76,6 +85,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `OutboundResponse` is now `ExecutionServiceOutboundResponseEvent`.
 - **BREAKING:**: Rename `MultichainRouter` to `MultichainRoutingService` and update action types accordingly ([#3913](https://github.com/MetaMask/snaps/pull/3913))
   - This is consistent with the naming of other services.
+- **BREAKING:** Rename `JsonSnapsRegistry` to `SnapsRegistryController` and update action types accordingly ([#3917](https://github.com/MetaMask/snaps/pull/3917))
+  - This is consistent with the naming of other controllers.
 - **BREAKING:** `MultichainRoutingService` now requires `SnapController:getRunnableSnaps` instead of `SnapController:getAllSnaps` ([#3913](https://github.com/MetaMask/snaps/pull/3913))
 - **BREAKING:** `SnapInsightsController` now requires `SnapController:getRunnableSnaps` instead of `SnapController:getAllSnaps` ([#3915](https://github.com/MetaMask/snaps/pull/3915))
 - **RREAKING:** Replace `ExecutionService` interface with abstract class ([#3916](https://github.com/MetaMask/snaps/pull/3916))